requirements.txt
gnews 
googlenewsdecoder 
newspaper3k 
lxml_html_clean 
tqdm
python-dotenv
google-generativeai
transformers 
torch
einops
groq